requirements.txt
alpaca-py
python-dotenv
ta